Splitting code in more specialised packages
