+++ src/main/java/de/zalando/ep/zalenium/proxy/CloudTestingRemoteProxy.java
@@ -10,7 +10,8 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import de.zalando.ep.zalenium.dashboard.Dashboard;
 import de.zalando.ep.zalenium.dashboard.TestInformation;
-import de.zalando.ep.zalenium.servlet.CloudProxyHtmlRenderer;
+import de.zalando.ep.zalenium.matcher.ZaleniumCapabilityMatcher;
+import de.zalando.ep.zalenium.servlet.renderer.CloudProxyHtmlRenderer;
 import de.zalando.ep.zalenium.util.*;
 import org.apache.commons.io.FileUtils;
 import org.openqa.grid.common.RegistrationRequest;
