+++ src/main/java/de/zalando/ep/zalenium/proxy/DockerSeleniumStarterRemoteProxy.java
@@ -5,7 +5,7 @@ import com.spotify.docker.client.DefaultDockerClient;
 import com.spotify.docker.client.DockerClient;
 import com.spotify.docker.client.exceptions.DockerException;
 import com.spotify.docker.client.messages.*;
-import de.zalando.ep.zalenium.util.DockerSeleniumCapabilityMatcher;
+import de.zalando.ep.zalenium.matcher.DockerSeleniumCapabilityMatcher;
 import de.zalando.ep.zalenium.util.Environment;
 import de.zalando.ep.zalenium.util.GoogleAnalyticsApi;
 import org.openqa.grid.common.RegistrationRequest;
