+++ src/main/java/de/zalando/ep/zalenium/servlet/LivePreviewServlet.java
@@ -8,6 +8,8 @@ package de.zalando.ep.zalenium.servlet;
 
 import com.google.common.io.ByteStreams;
 import de.zalando.ep.zalenium.proxy.DockerSeleniumRemoteProxy;
+import de.zalando.ep.zalenium.servlet.renderer.LiveNodeHtmlRenderer;
+import de.zalando.ep.zalenium.servlet.renderer.TemplateRenderer;
 import org.openqa.grid.internal.Registry;
 import org.openqa.grid.internal.RemoteProxy;
 import org.openqa.grid.internal.utils.HtmlRenderer;
